Improved user isolation. Users can no longer see eacother data by changing id in the url.

+++ web/Controllers/FriendshipsController.cs
@@ -36,6 +36,9 @@ namespace web.Controllers
         // GET: Friendships/Details/5
         public async Task<IActionResult> Details(string id)
         {
+
+            var currentUser = await _usermanager.GetUserAsync(User);
+
             if (id == null)
             {
                 return NotFound();
@@ -44,6 +47,7 @@ namespace web.Controllers
             var friendship = await _context.Friendships
                 .Include(f => f.User)
                 .Include(f => f.UserFriend)
+                .Where(f => f.User == currentUser)
                 .FirstOrDefaultAsync(m => m.UserId == id);
             if (friendship == null)
             {
@@ -92,7 +96,9 @@ namespace web.Controllers
             }
 
             var friendship = await _context.Friendships.FindAsync(id);
-            if (friendship == null)
+            var currentUser = await _usermanager.GetUserAsync(User);
+
+            if (friendship == null || friendship.User != currentUser)
             {
                 return NotFound();
             }
@@ -108,7 +114,8 @@ namespace web.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(string id, [Bind("UserId,UserFriendId")] Friendship friendship)
         {
-            if (id != friendship.UserId)
+            var currentUser = await _usermanager.GetUserAsync(User);
+            if (id != friendship.UserId || friendship.User != currentUser)
             {
                 return NotFound();
             }
@@ -146,10 +153,12 @@ namespace web.Controllers
                 return NotFound();
             }
 
+            var currentUser = await _usermanager.GetUserAsync(User);
+
             var friendship = await _context.Friendships
                 .Include(f => f.User)
                 .Include(f => f.UserFriend)
-                .FirstOrDefaultAsync(m => m.UserId == UserId && m.UserFriendId == UserFriendId);
+                .FirstOrDefaultAsync(m => m.UserId == UserId && m.UserFriendId == UserFriendId && m.User == currentUser);
             if (friendship == null)
             {
                 return NotFound();
@@ -163,7 +172,12 @@ namespace web.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string UserId, string UserFriendId)
         {
+            var currentUser = await _usermanager.GetUserAsync(User);
             var friendship = await _context.Friendships.FindAsync(UserId, UserFriendId);
+            if (friendship.User != currentUser)
+            {
+                return NotFound();
+            }
             _context.Friendships.Remove(friendship);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
